assignments/assignment-5/requirements.txt
# Core Dependencies
pydantic
python-dotenv
requests

# LangChain Ecosystem
langchain
langchain_core
langchain_community
langchain_google_genai

# Google Generative AI
google_generativeai

# Workflow Orchestration
langgraph

# Free API Dependencies
# OpenWeatherMap - Weather data
# Foursquare Places - Attractions, restaurants, activities
# ExchangeRate-API - Currency conversion
# No additional packages needed for these free APIs - using requests

# Testing Dependencies
pytest
pytest-cov
pytest-mock

# Additional Utilities
typing-extensions